+++ docs/history.md
@@ -6,8 +6,9 @@
 ### _v1.29.2_ - November 8th, 2023
 
 - Fixed bugs with the following babies:
-  - Abel (#585) - Tears that hit a poop or a fireplace no longer count as being missed.
+  - Driver Baby (#431) - Will no longer soft-lock in Downpour or Dross. (Thanks sisuka)
   - Cursed Pillow Baby (#487) - Tears that hit a poop or a fireplace no longer count as being missed. (Thanks sisuka)
+  - Abel (#585) - Tears that hit a poop or a fireplace no longer count as being missed.
 
 ### _v1.29.1_ - October 24th, 2023